+++ tests/Classes/PHP81.php
@@ -4,8 +4,11 @@ declare(strict_types=1);
 
 namespace Brick\Reflection\Tests\Classes;
 
+use ArrayAccess;
 use Brick\Reflection\Tests\Attributes\ExpectFunctionSignature;
+use Countable;
 use stdClass;
+use Traversable;
 
 abstract class PHP81
 {
@@ -14,4 +17,7 @@ abstract class PHP81
 
     #[ExpectFunctionSignature('abstract public function unionTypesWithFalse(string|false $a, int|false|null $b): \stdClass|false')]
     abstract public function unionTypesWithFalse(string|false $a, int|false|null $b): stdClass|false;
+
+    #[ExpectFunctionSignature('abstract public function intersectionTypes(\Countable&\Traversable $a): \Countable&\ArrayAccess')]
+    abstract public function intersectionTypes(Countable&Traversable $a): Countable&ArrayAccess;
 }